Move resource to separate directory
