+++ lunch_backend/resources/health.py
@@ -1,0 +1,6 @@
+from flask_restful import Resource
+
+
+class HealthCheck(Resource):
+    def get(self):
+        return {"status": "ok"}
